refactor(analyzer): modularize UI components and enhance style simulator architecture
This commit refactors the `:features:analyzer` module by extracting monolithic UI code into reusable, standalone components. It also completes the internationalization of the Style Simulator feature by replacing hardcoded strings with resource-based lookups.

- **`AnalyzerScreen.kt` & New UI Components**:
    - Decomposed the large `AnalyzerScreen` into modular components: `LocationInput`, `OccasionFilter`, `StyleCaptureSlot`, `StyleCaptureState`, and `AnalysisResultScreen`.
    - Improved code maintainability by isolating state-specific UI logic (Capture, Analysis, and Result) into dedicated files.
    - Updated `AnalysisResultScreen` to support direct navigation to "Nail Lab" and "Face Lab" experiences based on AI-generated suggestions.
- **Style Simulator Improvements**:
    - Introduced `SimulatorComponents.kt` to house specialized UI elements like `MagicBackground`, `MessagingStep`, and `ResultCard`.
    - Migrated all user-facing text in `StyleSimulatorScreen` and its components to `strings.xml`.
    - Refactored `ResultCard` to use `AsyncImage` for clothing item visualization and support dynamic color parsing.
- **`build.gradle.kts`**:
    - Added dependency on `:features:graphics` to utilize shared color utility functions and graphics processing.
- **`strings.xml`**:
    - Added comprehensive string resources for the Style Simulator, covering simulation steps (Bio-Markers, Rituals, Architecting) and result labels.
