+++ applications/kocolor/features/analyzer/src/main/res/values/strings.xml
@@ -32,4 +32,21 @@
     <string name="applications_kocolor_features_analyzer_experience">Experience</string>
     <string name="applications_kocolor_features_analyzer_discard">Discard</string>
     <string name="applications_kocolor_features_analyzer_save_analysis">Save Analysis</string>
+
+    <string name="applications_kocolor_features_analyzer_simulator_title">STYLE SIMULATOR</string>
+    <string name="applications_kocolor_features_analyzer_simulator_intent_title">Define Your Intent.</string>
+    <string name="applications_kocolor_features_analyzer_simulator_intent_placeholder">e.g. A crisp look for high-stakes negotiation.</string>
+    <string name="applications_kocolor_features_analyzer_simulator_begin_action">Begin Simulation</string>
+    <string name="applications_kocolor_features_analyzer_simulator_reading_bio">Reading Bio-Markers...</string>
+    <string name="applications_kocolor_features_analyzer_simulator_syncing_rituals">Syncing Daily Rituals...</string>
+    <string name="applications_kocolor_features_analyzer_simulator_architecting_style">Architecting Style...</string>
+    <string name="applications_kocolor_features_analyzer_simulator_magic">Magic...</string>
+    <string name="applications_kocolor_features_analyzer_simulator_synthesizing">Synthesizing your physical context with your stated intent.</string>
+    <string name="applications_kocolor_features_analyzer_simulator_blueprint">Your Blueprint.</string>
+    <string name="applications_kocolor_features_analyzer_simulator_local_calc">BEST-EFFORT LOCAL CALCULATION</string>
+    <string name="applications_kocolor_features_analyzer_simulator_ai_optimized">OPTIMIZED BY GEMINI AI</string>
+    <string name="applications_kocolor_features_analyzer_simulator_chromatic_core">CHROMATIC CORE</string>
+    <string name="applications_kocolor_features_analyzer_simulator_vault_selections">VAULT SELECTIONS</string>
+    <string name="applications_kocolor_features_analyzer_simulator_lock_palette">Lock Selection to Palette</string>
+    <string name="applications_kocolor_features_analyzer_simulator_bespoke">Bespoke</string>
 </resources>
